\name{germany}
\docType{data}
\alias{germany}
\title{germany}
\description{Data for Germen reunification.}
\format{dataframe}
\keyword{datasets}
\references{
  Abadie, R. et al (2014). Comparative Politics and the Synthetic Control Method. American Journal of Political Science 59(2): 495-510.
}

