\name{npc}
\docType{data}
\alias{npc}
\title{npc}
\description{Returns to Office of China's National People's Congress.}
\format{dataframe}
\keyword{datasets}
\references{
  Truex, R. (2014). The Returns to Office in a ``Rubber Stamp'' Parliament. American Political Science Review 108(2): 235--251.
}

