\name{print.tjbal}
\alias{print.tjbal}
\title{Print Results}
\description{Print results of trajectory balancing.}
\usage{\method{print}{tjbal}(x, \dots)} 
\arguments{
  \item{x}{a \code{\link{tjbal}} object.}
  \item{\dots}{other argv.}
}
\author{
  Chad Hazlett <chazlett@ucla.edu>, UCLA

  Yiqing Xu (Maintainer) <yiqingxu@ucsd.edu>, UCSD
}
\references{
  Hazlett, Chad and Yiqing Xu, 2018. ``Trajectory Balancing: A General Reweighting Approach to Causal Inference with Time-Series Cross-Sectional Data.'' Working Paper, UCLA and UCSD.
}
\seealso{
  \code{\link{tjbal}} and \code{\link{plot.tjbal}}
}


